chore: move schemas
