+++ src/apis/response-body-schema.ts
@@ -1,14 +1,14 @@
 import { z } from 'zod';
-import { ProgressSchema } from '../lib/types/schemas/progress-schema.ts';
+import { ProgressSchema } from '../lib/schemas/progress-schema.ts';
 import {
   CreatedRecruitmentSchema,
   RecruitmentSchema,
   RecruitmentSearchResultSchema,
-} from '../lib/types/schemas/recruitment-schema.ts';
+} from '../lib/schemas/recruitment-schema.ts';
 import {
   ApplicationDetailSchema,
   ApplicationOverviewSchema,
-} from '../lib/types/schemas/application-schema.ts';
+} from '../lib/schemas/application-schema.ts';
 
 const NoResponseDataSchema = z.literal('');
