+++ src/app/apply/[recruitmentCode]/_components/apply-narrative-box.tsx
@@ -4,7 +4,7 @@ import Typography from '../../../../components/atom/typography.tsx';
 import { IFormApplication } from '../page.tsx';
 import { cn } from '../../../../lib/utils/utils.ts';
 import { z } from 'zod';
-import { QuestionSchema } from '../../../../lib/types/schemas/question-schema.ts';
+import { QuestionSchema } from '../../../../lib/schemas/question-schema.ts';
 
 interface ApplyNarrativeBoxProps {
   question: z.infer<typeof QuestionSchema>;
