+++ src/app/apply/[recruitmentCode]/_components/apply-selective-box.tsx
@@ -6,7 +6,7 @@ import ApplyChoiceBox from './apply-choice-box';
 import { useState } from 'react';
 import { IFormApplication } from '../page.tsx';
 import { z } from 'zod';
-import { QuestionSchema } from '../../../../lib/types/schemas/question-schema.ts';
+import { QuestionSchema } from '../../../../lib/schemas/question-schema.ts';
 
 interface ApplySelectiveBoxProps {
   question: z.infer<typeof QuestionSchema>;
