+++ src/app/recruit/_components/recruit-make/details/question-boxes.tsx
@@ -19,10 +19,10 @@ import { useToast } from '../../../../../hooks/use-toast.ts';
 import {
   CREATED_NARRATIVE_QUESTION,
   CREATED_SELECTIVE_QUESTION,
-} from '../../../../../lib/types/default-data.ts';
+} from '../../../../../lib/schemas/default-data.ts';
 import { Button } from '../../../../../components/shadcn/button.tsx';
 import { z } from 'zod';
-import { CreatedRecruitmentSchema } from '../../../../../lib/types/schemas/recruitment-schema.ts';
+import { CreatedRecruitmentSchema } from '../../../../../lib/schemas/recruitment-schema.ts';
 
 const buttonDefaultStyle: string = 'h-4 w-4 cursor-pointer text-crews-g06';
