+++ src/app/sign-in/page.tsx
@@ -10,7 +10,7 @@ import {
   TooltipTrigger,
 } from '../../components/shadcn/tooltip.tsx';
 import { z } from 'zod';
-import { RoleSchema } from '../../lib/types/schemas/role-schema.ts';
+import { RoleSchema } from '../../lib/schemas/role-schema.ts';
 
 const Page = () => {
   const location = useLocation();
